lower plugin priority (#1)
At the moment, GME isn't used regardless of if we're probed after it, as
this plugin returns true for an is_our_file match, while GME always
returns false.

The lower priority means that vgmstream is now in the lead, getting rid
of the XA association issue I was having.

This may come back to bite me if Audacious' GME plugin ever implements
is_our_file, but for now this appears to work fine.

(For context, we try to avoid GME as its VGM player only supports the
Genesis and Master System, and plays blank audio if not those.)

This closes #1 (for now)